OAK-593: Segment-based MK
Use the same segment cache across file and MongoDB backends

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1525946 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-run/src/main/java/org/apache/jackrabbit/oak/benchmark/BenchmarkRunner.java
@@ -32,7 +32,7 @@ import org.apache.jackrabbit.oak.fixture.RepositoryFixture;
 
 public class BenchmarkRunner {
 
-    private static final long MB = 1024 * 1024;
+    private static final int MB = 1024 * 1024;
 
     public static void main(String[] args) throws Exception {
         OptionParser parser = new OptionParser();
